CleanUp Lock namespace - introduce enum Lock - "Type" - Move Lock/Lock to BaseLock
